requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
dataclasses-json==0.5.6
Django==3.2.7
django-activity-stream==0.10.0
django-crispy-forms==1.13.0
django-filter==21.1
django-taggit==1.5.1
djangorestframework==3.13.1
factory-boy==3.2.0
Faker==8.12.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
typing-extensions==4.0.1
typing-inspect==0.7.1
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
